feat: refactored annual review code
